modifiquei a classe Grafo para virar a classe Digrafo, a classe Grafo passa à ser uma subclasse de Digrafo
